refactor(charts): update LabCharts to direct imports from SimpleCharts/shared
SimpleCharts components import from the shared barrel themselves,
so the barrel is preserved. Follow-up spec needed to remove it.
