+++ app/src/components/Charts/LabCharts/Top10BillboardRace/Top10BillboardRaceView.tsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 import type { Top10BillboardRaceQueryResult } from './query'
 import type { EntityType } from '../shared/types'
 import { GhostLeaderboard, type GhostEntry } from './GhostLeaderboard'
-import { InsightCard } from '../../SimpleCharts/shared'
+import { InsightCard } from '../../SimpleCharts/shared/InsightCard'
 import { BAR_CHART_COLORS } from '../barChartColors'
 import { RaceControlBar } from '../Common/RaceControlBar'
 import { useRacePlayback } from '../Common/useRacePlayback'
